# openapi.json
{"openapi":"3.0.1","info":{"title":"GadgetBid API","description":"This is the core of Gadget Bid","license":{"name":"Gadget Bid Proprietary license","url":"https://www.gadgetbid.com"},"version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/user":{"post":{"tags":["user-controller"],"summary":"Create user","operationId":"userControllerCreateUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserInputDto"}}},"required":true},"responses":{"200":{"description":"Successful","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOutputDto"}}}},"401":{"description":"Unauthorized","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}},"404":{"description":"Not Found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}},"400":{"description":"Bad Request","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}}}},"/auth/updatePassword":{"post":{"tags":["auth-controller"],"operationId":"updatePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePasswordDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/auth/resetPassword":{"post":{"tags":["auth-controller"],"operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResetPasswordDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogInDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthSessionDto"}}}}}}},"/hello":{"get":{"tags":["test-controller"],"operationId":"test","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TestDTO"}}}}}}},"/hello/user":{"get":{"tags":["test-controller"],"operationId":"helloUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/hello/login":{"get":{"tags":["test-controller"],"operationId":"login_1","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/hello/admin":{"get":{"tags":["test-controller"],"operationId":"helloAdmin","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/auth/email/confirm":{"get":{"tags":["auth-controller"],"operationId":"confirmEmail","parameters":[{"name":"token","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/audit/list":{"get":{"tags":["audit-controller"],"operationId":"auditControllerGetAuditList","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/AuditDto"}}}}}}}}},"components":{"schemas":{"UserInputDto":{"required":["email","firstName","lastName","phone","role"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"type":"string"},"role":{"type":"string","enum":["ADMIN","ACCOUNT_MANAGER","USER"]}}},"UserAnswerOutputDto":{"required":["answer","createdAt","id","question","user"],"type":"object","properties":{"id":{"type":"number"},"user":{"$ref":"#/components/schemas/UserOutputDto"},"question":{"$ref":"#/components/schemas/UserQuestionDto"},"answer":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"UserKycFileDto":{"required":["createdAt","filePath","id","mediaId","sessionId","size","type","updatedAt","userId"],"type":"object","properties":{"id":{"type":"number"},"userId":{"type":"number"},"sessionId":{"type":"string"},"mediaId":{"type":"string"},"type":{"type":"string","enum":["IDENTITY","PHOTO"]},"filePath":{"type":"string"},"size":{"type":"number"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"UserOnboardFileOutputDto":{"required":["createdAt","fileExtension","filePath","id","mimeType","name","onboardId","size","status","type","updatedAt","userId"],"type":"object","properties":{"id":{"type":"number"},"onboardId":{"type":"number"},"userId":{"type":"number"},"type":{"type":"string","enum":["IDENTITY","INVESTMENT_AGREEMENT","INVESTMENT_AGREEMENT_SIGNED","SPV_AGREEMENT","SPV_AGREEMENT_SIGNED","PROOF_OF_PAYMENT","CONTRACT","CONTRACT_SIGNED","EXECUTIVE_SUMMARY"]},"name":{"type":"string"},"description":{"type":"string"},"mimeType":{"type":"string"},"fileExtension":{"type":"string"},"filePath":{"type":"string"},"size":{"type":"number"},"status":{"type":"string","enum":["NEW","APPROVED","DELETED"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"UserOnboardOutputDto":{"required":["accountManager","createdAt","files","id","status","updatedAt","user"],"type":"object","properties":{"id":{"type":"number"},"user":{"$ref":"#/components/schemas/UserOutputDto"},"files":{"type":"array","items":{"$ref":"#/components/schemas/UserOnboardFileOutputDto"}},"accountManager":{"$ref":"#/components/schemas/UserOutputDto"},"status":{"type":"string","enum":["OPEN","REGISTERED","NOT_VERIFIED","VERIFIED","KYC_KYB_AML","KYC_KYB_AML_OK","REJECTED","NOT_ELIGIBLE"]},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"UserOutputDto":{"type":"object","properties":{"id":{"type":"number"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"address":{"type":"string"},"personalNumber":{"type":"string"},"cardNumber":{"type":"string"},"walletId":{"type":"string"},"country":{"type":"string"},"dateOfBirth":{"type":"string"},"active":{"type":"number"},"role":{"type":"string","enum":["ADMIN","ACCOUNT_MANAGER","USER"]},"cardType":{"type":"string","enum":["PASSPORT","ID_CARD","DRIVERS_LICENSE","RESIDENCE_PERMIT","OTHER"]},"onboard":{"$ref":"#/components/schemas/UserOnboardOutputDto"},"accountManager":{"$ref":"#/components/schemas/UserOutputDto"},"kycCompleted":{"type":"number"},"tourComplete":{"type":"number"},"questionsCompleted":{"type":"number"},"verified":{"type":"number"},"twoFactorActivated":{"type":"number"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"kycFiles":{"type":"array","items":{"$ref":"#/components/schemas/UserKycFileDto"}},"answers":{"type":"array","items":{"$ref":"#/components/schemas/UserAnswerOutputDto"}},"qaScore":{"type":"number"}}},"UserQuestionDto":{"required":["active","id","order","question","variant1","variant2","variant3"],"type":"object","properties":{"id":{"type":"number"},"question":{"type":"string"},"variant1":{"type":"string"},"variant2":{"type":"string"},"variant3":{"type":"string"},"order":{"type":"number"},"active":{"type":"number"}}},"ErrorDto":{"required":["error","message","statusCode"],"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"array","items":{"type":"string"}},"error":{"type":"string"}}},"UpdatePasswordDto":{"required":["password","passwordConfirmation","token"],"type":"object","properties":{"password":{"maxLength":2147483647,"minLength":8,"type":"string"},"passwordConfirmation":{"type":"string"},"token":{"type":"string"}}},"ResetPasswordDto":{"required":["email"],"type":"object","properties":{"email":{"type":"string"}}},"RegisterDto":{"required":["email","firstName","lastName","password","passwordConfirmation","phone"],"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"phone":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"},"passwordConfirmation":{"maxLength":2147483647,"minLength":8,"type":"string"},"role":{"type":"string","enum":["ISSUER","INVESTOR"]}}},"LogInDto":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":8,"type":"string"}}},"AuthSessionDto":{"type":"object","properties":{"expiresIn":{"type":"string"},"isSecondFactorActive":{"type":"boolean"},"accessToken":{"type":"string"}}},"TestDTO":{"type":"object","properties":{"message":{"type":"string"}}},"AuditDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"url":{"type":"string"},"method":{"type":"string"},"payload":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}}},"securitySchemes":{"gadget-secured-api":{"type":"apiKey","name":"Authentication","in":"header"}}}}