# src/main/java/com/gadget/dto/audit/AuditDto.java
package com.gadget.dto.audit;

import com.fasterxml.jackson.databind.annotation.JsonSerialize;
import com.fasterxml.jackson.datatype.jsr310.ser.LocalDateTimeSerializer;
import lombok.Data;

import java.time.LocalDateTime;

@Data
public class AuditDto {

    private Long id;
    private String url;
    private String method;
    private String payload;
    @JsonSerialize(using = LocalDateTimeSerializer.class)
    private LocalDateTime createdAt;

}